+++ app_worker/ytdl_opts/per_host/__init__.py
@@ -1,7 +1,15 @@
 from ytdl_opts.per_host._default import DefaultHost
 from ytdl_opts.per_host.facebook import FacebookHost
 from ytdl_opts.per_host.instagram import InstagramHost
+from ytdl_opts.per_host.kika import KikaHost
 from ytdl_opts.per_host.tiktok import TikTokHost
 from ytdl_opts.per_host.twitter import TwitterHost
 
-__all__ = ['DefaultHost', 'FacebookHost', 'InstagramHost', 'TikTokHost', 'TwitterHost']
+__all__ = [
+    'DefaultHost',
+    'FacebookHost',
+    'InstagramHost',
+    'KikaHost',
+    'TikTokHost',
+    'TwitterHost',
+]