+++ yt_shared/src/yt_shared/constants.py
@@ -19,6 +19,7 @@ TWITTER_HOSTS: Final[tuple[str, ...]] = (
     'www.t.co',
 )
 FACEBOOK_HOSTS: Final[tuple[str, ...]] = ('facebook.com', 'www.facebook.com')
+KIKA_HOSTS: Final[tuple[str, ...]] = ('kika.de', 'www.kika.de')
 
 REMOVE_QUERY_PARAMS_HOSTS: Final[set[str]] = {
     *TWITTER_HOSTS,